# Conflicts:
#	code/__DEFINES/~~bubber_defines/role_preferences.dm
#	icons/map_icons/clothing/head/_head.dmi
#	tgstation.dme

+++ code/datums/elements/dangerous_surgical_removal.dm
@@ -10,13 +10,16 @@
 
 	/// whether the removal needs to be surgical for it to explode. If you're adding more modes, just pass the signal directly instead
 	var/surgical
+	/// whether the removal causes the organ to qdel
+	var/annihilate
 
-/datum/element/dangerous_organ_removal/Attach(datum/target, surgical = FALSE)
+/datum/element/dangerous_organ_removal/Attach(datum/target, surgical = FALSE, annihilate = TRUE) //BUBBER EDIT - Adds new variable to make organ not qdel
 	. = ..()
 	if(!isorgan(target))
 		return ELEMENT_INCOMPATIBLE
 
 	src.surgical = surgical
+	src.annihilate = annihilate
 
 	if(surgical)
 		RegisterSignal(target, COMSIG_ORGAN_SURGICALLY_REMOVED, PROC_REF(on_removal))
@@ -40,4 +43,10 @@
 		source.audible_message("[source] explodes into tiny pieces!")
 
 	explosion(source, light_impact_range = 1, explosion_cause = source)
-	qdel(source)
+	//BUBBER EDIT - if it doesn't qdel, set it to take all its health in damage
+	//qdel(source)
+	if(annihilate)
+		qdel(source)
+	else
+		source.set_organ_damage(source.maxHealth)
+	//EDIT END
